add boot.crypto-early functionality
currently boot.crypto has run after boot.localfs for the reason
that file system conainers have to be available in localfs
before the crypto device could be setup.
Another valid use case however is to encrypt whole block devices
and use them for LVM or MD. In order to get this working the
boot.crypto script has to be invoked twice:
1. before boot.lvm and boot.md as boot.crypto-early
2. after boot.localfs
a new option for crypttab "noearly" is introduced to tag
devices which should be skipped in the first invocation
(this would affect the crypto file container).